+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/train/adapter/in/TrainController.kt
@@ -4,7 +4,6 @@ import backend.team.ahachul_backend.api.train.adapter.`in`.dto.GetCongestionDto
 import backend.team.ahachul_backend.api.train.adapter.`in`.dto.GetTrainDto
 import backend.team.ahachul_backend.api.train.adapter.`in`.dto.GetTrainRealTimesDto
 import backend.team.ahachul_backend.api.train.application.port.`in`.TrainUseCase
-import backend.team.ahachul_backend.api.train.application.service.TrainCongestionService
 import backend.team.ahachul_backend.common.annotation.Authentication
 import backend.team.ahachul_backend.common.response.CommonResponse
 import org.springframework.web.bind.annotation.GetMapping
@@ -13,8 +12,7 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 class TrainController(
-    private val trainUseCase: TrainUseCase,
-    private val trainCongestionService: TrainCongestionService
+    private val trainUseCase: TrainUseCase
 ) {
 
     @Authentication
@@ -32,7 +30,7 @@ class TrainController(
     @Authentication
     @GetMapping("/v1/trains/real-times/congestion/{stationId}")
     fun getCongestion(@PathVariable("stationId") stationId: Long): CommonResponse<GetCongestionDto.Response> {
-        val result = trainCongestionService.getTrainCongestion(stationId)
+        val result = trainUseCase.getTrainCongestion(stationId)
         return CommonResponse.success(result)
     }
 }
