+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/train/application/port/in/TrainUseCase.kt
@@ -1,5 +1,6 @@
 package backend.team.ahachul_backend.api.train.application.port.`in`
 
+import backend.team.ahachul_backend.api.train.adapter.`in`.dto.GetCongestionDto
 import backend.team.ahachul_backend.api.train.adapter.`in`.dto.GetTrainDto
 import backend.team.ahachul_backend.api.train.adapter.`in`.dto.GetTrainRealTimesDto
 
@@ -8,4 +9,6 @@ interface TrainUseCase {
     fun getTrain(trainNo: String): GetTrainDto.Response
 
     fun getTrainRealTimes(stationId: Long): List<GetTrainRealTimesDto.TrainRealTime>
+
+    fun getTrainCongestion(stationId: Long): GetCongestionDto.Response
 }
